backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.4
numpy==1.26.2
websockets==12.0
aiofiles==23.2.1
# Better LLM models
llama-cpp-python  # For GGUF model inference
sentence-transformers  # For embeddings
chromadb  # Vector database for RAG
# Qwen + RAG Architecture
llama-cpp-python==0.2.90
chromadb==0.4.24
sentence-transformers==2.7.0
langchain==0.2.0
langchain-community==0.2.0
flagembedding==1.2.0